Extract XlsxScenarioExporter from ProjectScenarioFileExport (1207-line class → focused exporter + facade)
